first commit. change the file organisation and bug correction
